refactor(core): drop dead cli duplicates, split extractors
