--- packages/core/src/summarizer/cli.ts
@@ -1,324 +1,0 @@
-import { spawn } from 'node:child_process';
-import type { ContextSlot, ProviderId } from '@goodboy/types';
-import { isSlotKey, SLOT_KEYS, type SlotKey } from '../context/slots';
-import { computeCostUsd } from '../providers/claude/cost';
-import { PROVIDER_CAPABILITIES } from '../providers/capabilities';
-import { CURSOR_AUTO_MODEL } from '../providers/cursor/models';
-import { inferNextActions, type NextAction, type NextActionsPrState } from './next-actions';
-import { SUMMARIZER_SYSTEM_PROMPT } from './prompt';
-
-type ContextSlotDeltaUpsert = Readonly<{ key: SlotKey; value: string }>;
-
-type ContextSlotDelta = Readonly<{
-  upserts: ReadonlyArray<ContextSlotDeltaUpsert>;
-}>;
-
-export type SummarizerUsage = {
-  readonly inputTokens: number;
-  readonly outputTokens: number;
-  readonly cachedInputTokens: number;
-  readonly estimatedCostUsd: number;
-};
-
-export type SummarizeInput = {
-  readonly prevSlots: ReadonlyArray<ContextSlot>;
-  readonly turnInput: string;
-  readonly turnOutput: string;
-  readonly prState?: NextActionsPrState | null;
-};
-
-export type SummarizerResult = {
-  readonly delta: ContextSlotDelta;
-  readonly usage: SummarizerUsage;
-  readonly model: string;
-  readonly nextActions: ReadonlyArray<NextAction>;
-};
-
-export type SummarizerDeps = {
-  readonly providerId: ProviderId;
-  readonly binary?: string;
-  readonly spawnFn?: typeof spawn;
-};
-
-export class SummarizerSpawnError extends Error {
-  constructor(
-    public readonly exitCode: number | null,
-    public readonly stderr: string,
-  ) {
-    super(`summarizer cli exited with code ${exitCode ?? 'null'}`);
-    this.name = 'SummarizerSpawnError';
-  }
-}
-
-export class SummarizerParseError extends Error {
-  constructor(
-    message: string,
-    public readonly raw: string,
-  ) {
-    super(message);
-    this.name = 'SummarizerParseError';
-  }
-}
-
-function getCheapModel(providerId: ProviderId): string {
-  if (providerId === 'cursor') {
-    return CURSOR_AUTO_MODEL;
-  }
-  const caps = PROVIDER_CAPABILITIES[providerId];
-  return caps.models.find((m) => m.tier === 'cheap')?.id ?? caps.models[0]!.id;
-}
-
-function getDefaultBinary(providerId: ProviderId): string {
-  switch (providerId) {
-    case 'anthropic':
-      return 'claude';
-    case 'cursor':
-      return 'cursor-agent';
-    case 'codex':
-      return 'codex';
-    case 'gemini':
-      return 'gemini';
-    default: {
-      const _exhaustive: never = providerId;
-      throw new Error(`unknown provider: ${_exhaustive}`);
-    }
-  }
-}
-
-type ClaudeJsonResult = {
-  readonly result?: string;
-  readonly usage?: {
-    readonly input_tokens?: number;
-    readonly output_tokens?: number;
-    readonly cache_read_input_tokens?: number;
-  };
-  readonly model?: string;
-  readonly subtype?: string;
-  readonly is_error?: boolean;
-};
-
-export class Summarizer {
-  private readonly providerId: ProviderId;
-  private readonly binary: string;
-  private readonly model: string;
-  private readonly spawnFn: typeof spawn;
-
-  constructor(deps: SummarizerDeps) {
-    this.providerId = deps.providerId;
-    this.binary = deps.binary ?? getDefaultBinary(deps.providerId);
-    this.model = getCheapModel(deps.providerId);
-    this.spawnFn = deps.spawnFn ?? spawn;
-  }
-
-  async summarize(input: SummarizeInput): Promise<SummarizerResult> {
-    const userMessage = buildUserPrompt(input);
-    const { stdout, stderr, exitCode } = await this.spawnCli(userMessage);
-
-    if (exitCode !== 0) {
-      throw new SummarizerSpawnError(exitCode, stderr);
-    }
-
-    const { text, usage } = this.extractTextAndUsage(stdout);
-    const delta = parseDelta(text);
-    const slotsAfter = applyDelta(input.prevSlots, delta);
-    const nextActions = inferNextActions({
-      input,
-      delta,
-      slotsAfter,
-      prState: input.prState ?? null,
-    });
-
-    return { delta, usage, model: this.model, nextActions };
-  }
-
-  private spawnCli(
-    userMessage: string,
-  ): Promise<{ stdout: string; stderr: string; exitCode: number | null }> {
-    return new Promise((resolve, reject) => {
-      const args = buildCliArgs(this.providerId, this.model, SUMMARIZER_SYSTEM_PROMPT, userMessage);
-      const child = this.spawnFn(this.binary, args, {
-        stdio: ['ignore', 'pipe', 'pipe'],
-      });
-
-      let stdout = '';
-      let stderr = '';
-
-      child.stdout?.on('data', (chunk: Buffer) => {
-        stdout += chunk.toString('utf8');
-      });
-      child.stderr?.on('data', (chunk: Buffer) => {
-        stderr += chunk.toString('utf8');
-      });
-      child.on('error', (err) => reject(err));
-      child.on('close', (code) => resolve({ stdout, stderr, exitCode: code }));
-    });
-  }
-
-  private extractTextAndUsage(stdout: string): { text: string; usage: SummarizerUsage } {
-    if (this.providerId === 'anthropic') {
-      return extractClaudeJsonOutput(stdout, this.model);
-    }
-    return { text: stdout.trim(), usage: zeroUsage() };
-  }
-}
-
-function buildCliArgs(
-  providerId: ProviderId,
-  model: string,
-  systemPrompt: string,
-  userMessage: string,
-): string[] {
-  switch (providerId) {
-    case 'anthropic':
-      return [
-        '-p',
-        userMessage,
-        '--model',
-        model,
-        '--system-prompt',
-        systemPrompt,
-        '--output-format',
-        'json',
-        '--no-session-persistence',
-      ];
-    case 'cursor':
-      return [
-        '-p',
-        `${systemPrompt}\n\n${userMessage}`,
-        '--model',
-        model,
-        '--output-format',
-        'stream-json',
-        '--force',
-      ];
-    case 'codex':
-      return [
-        'exec',
-        '--json',
-        '-m',
-        model,
-        '-s',
-        'read-only',
-        '--skip-git-repo-check',
-        `${systemPrompt}\n\n${userMessage}`,
-      ];
-    case 'gemini':
-      return ['-m', model, '-p', `${systemPrompt}\n\n${userMessage}`];
-    default: {
-      const _exhaustive: never = providerId;
-      throw new Error(`unknown provider: ${_exhaustive}`);
-    }
-  }
-}
-
-function extractClaudeJsonOutput(
-  stdout: string,
-  model: string,
-): { text: string; usage: SummarizerUsage } {
-  const trimmed = stdout.trim();
-  let parsed: ClaudeJsonResult;
-  try {
-    parsed = JSON.parse(trimmed) as ClaudeJsonResult;
-  } catch {
-    return { text: trimmed, usage: zeroUsage() };
-  }
-
-  const text = parsed.result ?? '';
-  const rawUsage = parsed.usage ?? {};
-  const inputTokens = rawUsage.input_tokens ?? 0;
-  const outputTokens = rawUsage.output_tokens ?? 0;
-  const cachedInputTokens = rawUsage.cache_read_input_tokens ?? 0;
-  const estimatedCostUsd = computeCostUsd(
-    { inputTokens, outputTokens, cachedInputTokens, estimatedCostUsd: 0 },
-    model,
-  );
-  return { text, usage: { inputTokens, outputTokens, cachedInputTokens, estimatedCostUsd } };
-}
-
-function zeroUsage(): SummarizerUsage {
-  return { inputTokens: 0, outputTokens: 0, cachedInputTokens: 0, estimatedCostUsd: 0 };
-}
-
-function buildUserPrompt(input: SummarizeInput): string {
-  const slotLines = SLOT_KEYS.map((key) => {
-    const slot = input.prevSlots.find((s) => s.key === key);
-    const value = slot?.enabled ? slot.value || '(empty)' : '(empty)';
-    return `${key}: ${value}`;
-  }).join('\n');
-
-  return [
-    'Current slot values:',
-    slotLines,
-    '',
-    'User turn:',
-    input.turnInput,
-    '',
-    'Assistant turn:',
-    input.turnOutput,
-    '',
-    'Return the JSON object now.',
-  ].join('\n');
-}
-
-function parseDelta(raw: string): ContextSlotDelta {
-  const stripped = stripCodeFences(raw);
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(stripped);
-  } catch (err) {
-    throw new SummarizerParseError(
-      `summarizer response was not valid JSON: ${err instanceof Error ? err.message : String(err)}`,
-      raw,
-    );
-  }
-
-  if (typeof parsed !== 'object' || parsed === null || !('upserts' in parsed)) {
-    throw new SummarizerParseError('summarizer response missing "upserts" array', raw);
-  }
-  const candidate = (parsed as { upserts: unknown }).upserts;
-  if (!Array.isArray(candidate)) {
-    throw new SummarizerParseError('summarizer "upserts" was not an array', raw);
-  }
-
-  const upserts: ContextSlotDeltaUpsert[] = [];
-  for (const entry of candidate) {
-    if (typeof entry !== 'object' || entry === null) {
-      continue;
-    }
-    const e = entry as Record<string, unknown>;
-    const key = e.key;
-    const value = e.value;
-    if (typeof key !== 'string' || !isSlotKey(key)) {
-      continue;
-    }
-    if (typeof value !== 'string') {
-      continue;
-    }
-    upserts.push({ key, value });
-  }
-  return { upserts };
-}
-
-function stripCodeFences(raw: string): string {
-  const fenced = /^```(?:json)?\s*([\s\S]*?)\s*```$/i.exec(raw.trim());
-  return (fenced?.[1] ?? raw).trim();
-}
-
-function applyDelta(
-  prev: ReadonlyArray<ContextSlot>,
-  delta: ContextSlotDelta,
-): ReadonlyArray<ContextSlot> {
-  if (delta.upserts.length === 0) {
-    return prev;
-  }
-  const byKey = new Map<string, ContextSlot>(prev.map((s) => [s.key, s]));
-  for (const upsert of delta.upserts) {
-    const existing = byKey.get(upsert.key);
-    byKey.set(upsert.key, {
-      key: upsert.key,
-      value: upsert.value,
-      enabled: existing?.enabled ?? true,
-    });
-  }
-  return Array.from(byKey.values());
-}
